move create menu to component for page speed, add todos

+++ pages/app.vue
@@ -18,140 +18,30 @@
       </ul>
     </nav>
 
-    <span v-if="filteredLifts.length === 0" class="block text-center"
+    <span
+      v-if="filteredLifts.length === 0"
+      class="block text-center text-gray-700"
       >No workouts for {{ currentCategory }}</span
     >
 
     <template v-for="lift in filteredLifts">
-      <WorkoutCard :lift="lift" @deleteLift="deleteLift(lift.id)" />
+      <WorkoutCard :lift="lift" @delete-lift="deleteLift(lift.id)" />
     </template>
 
     <USlideover :modelValue="addMenuOpen">
-      <div class="flex justify-start pl-2 pt-2">
-        <UButton
-          @click="addMenuOpen = false"
-          variant="link"
-          icon="i-heroicons-arrow-right"
-        ></UButton>
-      </div>
-      <h2 class="px-4 font-bold text-4xl mt-12 mb-6">Add A Lift:</h2>
-      <form class="flex flex-col my-2 px-4" @submit.prevent="addLift">
-        <div>
-          <label for="name" class="font-bold">Lift Name:</label>
-          <UInput
-            v-model="newLift.name"
-            class="w-full bg-inherit dark:bg-inherit mb-4"
-            size="xl"
-            type="text"
-            color="primary"
-            name="name"
-            autofocus
-            required
-            autocomplete="off"
-          />
-
-          <label for="category" class="font-bold">Category:</label>
-          <USelect
-            v-model="newLift.category"
-            size="xl"
-            :options="categories"
-            color="primary"
-            class="w-full mb-4"
-            name="category"
-            placeholder="Select Category"
-            required
-          >
-          </USelect>
+      <AddLiftMenu
+        :newLift="newLift"
+        :categories="categories"
+        @add-lift="addLift"
+      >
+        <div class="flex justify-start pl-2 pt-2">
+          <UButton
+            @click="addMenuOpen = false"
+            variant="link"
+            icon="i-heroicons-arrow-right"
+          ></UButton>
         </div>
-
-        <div v-if="newLift.category !== 'Cardio'">
-          <label for="weight" class="font-bold">Weight:</label>
-          <UInput
-            v-model="newLift.weight"
-            class="w-full bg-inherit dark:bg-inherit mb-4"
-            size="xl"
-            type="number"
-            color="primary"
-            name="weight"
-            autocomplete="off"
-          />
-
-          <label for="reps" class="font-bold">Reps:</label>
-          <div class="flex gap-2">
-            <UInput
-              v-model="newLift.reps[0]"
-              class="w-full bg-inherit dark:bg-inherit mb-4"
-              size="xl"
-              type="number"
-              color="primary"
-              name="reps"
-              autocomplete="off"
-            />
-            <UInput
-              v-model="newLift.reps[1]"
-              class="w-full bg-inherit dark:bg-inherit mb-4"
-              size="xl"
-              type="number"
-              color="primary"
-              autocomplete="off"
-            />
-            <UInput
-              v-model="newLift.reps[2]"
-              class="w-full bg-inherit dark:bg-inherit mb-4"
-              size="xl"
-              type="number"
-              color="primary"
-              autocomplete="off"
-            />
-            <UInput
-              v-model="newLift.reps[3]"
-              class="w-full bg-inherit dark:bg-inherit mb-4"
-              size="xl"
-              type="number"
-              color="primary"
-              autocomplete="off"
-            />
-            <UInput
-              v-model="newLift.reps[4]"
-              class="w-full bg-inherit dark:bg-inherit mb-4"
-              size="xl"
-              type="number"
-              color="primary"
-              autocomplete="off"
-            />
-          </div>
-        </div>
-
-        <div v-else>
-          <label for="time" class="font-bold">Time:</label>
-          <UInput
-            v-model="newLift.time"
-            class="w-full bg-inherit dark:bg-inherit mb-4"
-            size="xl"
-            type="number"
-            color="primary"
-            name="time"
-            autocomplete="off"
-          />
-
-          <label for="speed" class="font-bold">Speed/Intensity:</label>
-          <UInput
-            v-model="newLift.speed"
-            class="w-full bg-inherit dark:bg-inherit mb-4"
-            size="xl"
-            type="number"
-            color="primary"
-            name="speed"
-            autocomplete="off"
-          />
-        </div>
-
-        <div class="flex place-self-end w-1/2 mt-2">
-          <UButton type="submit" block size="xl" :loading="loading">
-            Add
-          </UButton>
-        </div>
-      </form>
+      </AddLiftMenu>
     </USlideover>
 
     <USlideover :modelValue="timerMenuOpen" side="left">
@@ -190,6 +80,7 @@
 </template>
 
 <script setup lang="ts">
+import { registerRuntimeCompiler } from "nuxt/dist/app/compat/capi";
 import { Database } from "~~/types/database.types";
 import { Lift, CategoryEnums } from "~~/types/lifts";
 
@@ -289,6 +180,9 @@ async function addLift() {
 // }
 
 const deleteLift = async (liftId: number) => {
+  const prompt = confirm("Delete this lift forever?");
+  if (!prompt) return;
+
   await client.from("lifts").delete().match({ id: liftId });
 
   lifts.value = lifts.value.filter((lift: Lift) => lift.id !== liftId);
